# week2-api-design-management/api-swagger/src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: 이커머스 서비스 API
  description: "이커머스 서비스의 RESTful API 명세서입니다. 상품 조회, 장바구니 관리, 주문/결제, 쿠폰 관리 기능을 제공합\
    니다."
  version: v1
servers:
- url: http://localhost:8080
  description: 로컬 개발 서버
...
paths:
  /api/v1/orders:
    get:
      tags:
      - 주문 API
      summary: 주문 내역 조회
      description: 사용자의 주문 내역을 페이지네이션하여 조회합니다.
      operationId: getOrders
      parameters:
      - name: userId
        in: query
        description: 사용자 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      - name: page
        in: query
        description: 페이지 번호 (0부터 시작)
        required: false
        schema:
          type: integer
          format: int32
          default: 0
        example: 0
      - name: size
        in: query
        description: 페이지 크기
        required: false
        schema:
          type: integer
          format: int32
          default: 10
        example: 10
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
    post:
      tags:
      - 주문 API
      summary: 주문 생성
      description: 장바구니의 상품으로 주문을 생성하고 결제를 처리합니다. 쿠폰이 있는 경우 할인이 적용됩니다.
      operationId: createOrder
      requestBody:
        description: 주문 생성 요청 정보
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
        required: true
      responses:
        "400":
          description: "잘못된 요청 (장바구니 비어있음, 잔액 부족, 쿠폰 이미 사용됨, 쿠폰 만료)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: 재고 부족
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "201":
          description: 주문 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        "404":
          description: 쿠폰을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/cart/items:
    post:
      tags:
      - 장바구니 API
      summary: 장바구니에 상품 추가
      description: 장바구니에 상품을 추가하거나 기존 수량을 증가시킵니다.
      operationId: addCartItem
      requestBody:
        description: 장바구니 상품 추가 요청
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCartItemRequest'
        required: true
      responses:
        "409":
          description: 재고 부족
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: 상품을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "201":
          description: 추가 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddCartItemResponse'
...
  /api/v1/cart/items/{cartItemId}:
    delete:
      tags:
      - 장바구니 API
      summary: 장바구니 상품 삭제
      description: 장바구니에서 특정 상품을 제거합니다.
      operationId: deleteCartItem
      parameters:
      - name: cartItemId
        in: path
        description: 장바구니 항목 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "204":
          description: 삭제 성공 (응답 본문 없음)
        "404":
          description: 장바구니 항목을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
      - 장바구니 API
      summary: 장바구니 상품 수량 변경
      description: 장바구니에 담긴 상품의 수량을 변경합니다.
      operationId: updateCartItem
      parameters:
      - name: cartItemId
        in: path
        description: 장바구니 항목 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        description: 장바구니 상품 수량 변경 요청
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartItemRequest'
        required: true
      responses:
        "409":
          description: 재고 부족
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: 변경 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateCartItemResponse'
        "404":
          description: 장바구니 항목을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/orders/{orderId}:
    get:
      tags:
      - 주문 API
      summary: 주문 상세 조회
      description: "특정 주문의 상세 정보를 조회합니다. 주문 상품 목록, 쿠폰 사용 정보, 결제 정보를 포함합니다."
      operationId: getOrder
      parameters:
      - name: orderId
        in: path
        description: 주문 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 12345
      responses:
        "404":
          description: 주문을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
...
  /api/v1/cart:
    get:
      tags:
      - 장바구니 API
      summary: 장바구니 조회
      description: 사용자의 장바구니 내역을 조회합니다.
      operationId: getCart
      parameters:
      - name: userId
        in: query
        description: 사용자 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
...
components:
  schemas:
    CreateOrderRequest:
      required:
      - userId
      type: object
      properties:
        userId:
          type: integer
          description: 사용자 ID
          format: int64
          example: 1
        couponId:
          type: integer
          description: 쿠폰 ID (없으면 할인 미적용)
          format: int64
          example: 5
      description: 주문 생성 요청
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
          description: 에러 코드
          example: USER_NOT_FOUND
        message:
          type: string
          description: 에러 메시지
          example: 사용자를 찾을 수 없습니다.
        details:
          type: string
          description: 에러 상세 설명 (optional)
          example: "userId: 1"
      description: 에러 상세 정보
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
        timestamp:
          type: string
          description: 에러 발생 시간
          format: date-time
      description: 에러 응답
    CouponUsed:
      type: object
      properties:
        couponId:
          type: integer
          description: 쿠폰 ID
          format: int64
          example: 5
        couponName:
          type: string
          description: 쿠폰명
          example: 신규 가입 쿠폰
        discountAmount:
          type: integer
          description: 할인 금액
          format: int64
          example: 100000
      description: 사용된 쿠폰 정보
    OrderItem:
      type: object
      properties:
        orderItemId:
          type: integer
          description: 주문 상품 ID
          format: int64
          example: 1
        productId:
          type: integer
          description: 상품 ID
          format: int64
          example: 1
        productName:
          type: string
          description: 상품명
          example: 노트북
        price:
          type: integer
          description: 상품 단가
          format: int64
          example: 1500000
        quantity:
          type: integer
          description: 수량
          format: int32
          example: 2
        subtotal:
          type: integer
          description: 소계 (단가 x 수량)
          format: int64
          example: 3000000
      description: 주문 상품 정보
    OrderResponse:
      type: object
      properties:
        orderId:
          type: integer
          description: 주문 ID
          format: int64
          example: 12345
        userId:
          type: integer
          description: 사용자 ID
          format: int64
          example: 1
        status:
          type: string
          description: 주문 상태
          example: COMPLETED
        items:
          type: array
          description: 주문 상품 목록
          items:
            $ref: '#/components/schemas/OrderItem'
        originalAmount:
          type: integer
          description: 원래 금액
          format: int64
          example: 3000000
        discountAmount:
          type: integer
          description: 할인 금액
          format: int64
          example: 100000
        finalAmount:
          type: integer
          description: 최종 결제 금액
          format: int64
          example: 2900000
        couponUsed:
          $ref: '#/components/schemas/CouponUsed'
        paymentInfo:
          $ref: '#/components/schemas/PaymentInfo'
        createdAt:
          type: string
          description: 주문 생성 시간
          format: date-time
        updatedAt:
          type: string
          description: 주문 수정 시간
          format: date-time
      description: 주문 상세 응답
    PaymentInfo:
      type: object
      properties:
        paymentId:
          type: integer
          description: 결제 ID
          format: int64
          example: 678
        method:
          type: string
          description: 결제 수단
          example: BALANCE
        amount:
          type: integer
          description: 결제 금액
          format: int64
          example: 2900000
        paidAt:
          type: string
          description: 결제 완료 시간
          format: date-time
      description: 결제 정보
...
    AddCartItemRequest:
      required:
      - productId
      - quantity
      - userId
      type: object
      properties:
        userId:
          minimum: 1
          type: integer
          description: 사용자 ID
          format: int64
          example: 1
        productId:
          minimum: 1
          type: integer
          description: 상품 ID
          format: int64
          example: 1
        quantity:
          minimum: 1
          type: integer
          description: 수량
          format: int32
          example: 2
      description: 장바구니 상품 추가 요청
    AddCartItemResponse:
      type: object
      properties:
        cartItemId:
          type: integer
          description: 장바구니 항목 ID
          format: int64
          example: 1
        productId:
          type: integer
          description: 상품 ID
          format: int64
          example: 1
        productName:
          type: string
          description: 상품명
          example: 노트북
        price:
          type: integer
          description: 상품 가격
          format: int64
          example: 1500000
        quantity:
          type: integer
          description: 수량
          format: int32
          example: 2
        subtotal:
          type: integer
          description: 소계
          format: int64
          example: 3000000
      description: 장바구니 상품 추가 응답
...
    UpdateCartItemRequest:
      required:
      - quantity
      type: object
      properties:
        quantity:
          minimum: 1
          type: integer
          description: 변경할 수량
          format: int32
          example: 5
      description: 장바구니 상품 수량 변경 요청
    UpdateCartItemResponse:
      type: object
      properties:
        cartItemId:
          type: integer
          description: 장바구니 항목 ID
          format: int64
          example: 1
        productId:
          type: integer
          description: 상품 ID
          format: int64
          example: 1
        productName:
          type: string
          description: 상품명
          example: 노트북
        price:
          type: integer
          description: 상품 가격
          format: int64
          example: 1500000
        quantity:
          type: integer
          description: 변경된 수량
          format: int32
          example: 5
        subtotal:
          type: integer
          description: 소계
          format: int64
          example: 7500000
      description: 장바구니 상품 수량 변경 응답
    PaginationInfo:
      type: object
      properties:
        currentPage:
          type: integer
          description: 현재 페이지 번호
          format: int32
          example: 0
        totalPages:
          type: integer
          description: 전체 페이지 수
          format: int32
          example: 5
        totalElements:
          type: integer
          description: 전체 요소 수
          format: int64
          example: 100
        size:
          type: integer
          description: 페이지 크기
          format: int32
          example: 20
      description: 페이지네이션 정보
...
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          description: 주문 목록
          items:
            $ref: '#/components/schemas/OrderSummary'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
      description: 주문 목록 응답
    OrderSummary:
      type: object
      properties:
        orderId:
          type: integer
          description: 주문 ID
          format: int64
          example: 12345
        status:
          type: string
          description: 주문 상태
          example: COMPLETED
        originalAmount:
          type: integer
          description: 원래 금액
          format: int64
          example: 3000000
        discountAmount:
          type: integer
          description: 할인 금액
          format: int64
          example: 100000
        finalAmount:
          type: integer
          description: 최종 결제 금액
          format: int64
          example: 2900000
        itemCount:
          type: integer
          description: 주문 상품 개수
          format: int32
          example: 2
        createdAt:
          type: string
          description: 주문 생성 시간
          format: date-time
      description: 주문 요약 정보
...
    CartItem:
      type: object
      properties:
        cartItemId:
          type: integer
          description: 장바구니 항목 ID
          format: int64
          example: 1
        productId:
          type: integer
          description: 상품 ID
          format: int64
          example: 1
        productName:
          type: string
          description: 상품명
          example: 노트북
        price:
          type: integer
          description: 상품 가격
          format: int64
          example: 1500000
        quantity:
          type: integer
          description: 수량
          format: int32
          example: 2
        subtotal:
          type: integer
          description: 소계
          format: int64
          example: 3000000
        stock:
          type: integer
          description: 재고
          format: int32
          example: 50
      description: 장바구니 상품 정보
    CartResponse:
      type: object
      properties:
        userId:
          type: integer
          description: 사용자 ID
          format: int64
          example: 1
        items:
          type: array
          description: 장바구니 상품 목록
          items:
            $ref: '#/components/schemas/CartItem'
        totalAmount:
          type: integer
          description: 총 금액
          format: int64
          example: 3030000
        totalItems:
          type: integer
          description: 총 상품 개수
          format: int32
          example: 2
      description: 장바구니 조회 응답
...